requirements.txt
fastapi>=0.136.0
uvicorn>=0.46.0
httpx>=0.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12

# Development / testing
pytest>=9.0
pytest-asyncio>=1.0